requirements.txt
Django==1.11.6
django-ckeditor==5.4.0
django-crispy-forms==1.6.1
django-ipware==1.1.6
django-js-asset==0.1.1
django-lockdown==1.4.2
django-multiupload==0.5.2
gunicorn==19.7.1
Markdown==2.6.10
olefile==0.44
Pillow==4.3.0
pkg-resources==0.0.0
python-slugify==1.2.4
pytz==2017.2
six==1.11.0
transliterate==1.10
Unidecode==0.4.21
Unipath==1.1
